Added reusable storage.clock
